use sbt-pgp in a community-build-friendly way
dbuild uses `publish`, so we want to avoid having it interactively
prompt for PGP stuff.  I looked at how other projects that use sbt-pgp
are doing it, and they all seem to do it this way instead.

the sbt-pgp version upgrade might not actually be necessary here, but
seems good to include regardless