Add the paradoxTheme to libraryDependencies using a special scope
This shall prevent that `paradoxTheme` is added to the dependencies
of the project using paradox, which would mean that project users would
unexpectedly get the paradoxTheme as a transitive dependency.

The newly introduced scope required to nudge sbt-web to grab
the assets out of it instead of `Compile` configuration graph.

Fixes #225