Implemented markdown includes
Fixes #195

Implemented markdown includes. It's very similar to snips, except
that instead of outputting the content in a verbatim node that then
gets rendered, it parses the content with the reader and then writes it
with a writer.

A new context is created for the writer to allow the included markdown
file to include other markdown files and code snippets relative to
itself. However, not a lot else about the context is modified, for
example, custom properties for the included page aren't read. This might
be desirable in some cases, and indesirable in others, not sure, I think
it's best to wait and see what issues people report with it to see what
other context should be modified.