Improved includes
This improves the includes, so rather than the included markdown being
parsed at the parent parents render time, it's parsed and then included
in the AST straight after the parent page is parsed.

One consequence of this is that now the left hand page navigation works,
since the headers can be extracted from the included markdown. I think
it's also more robust all around.

In order to implement this, I had to create a new node, an IncludeNode,
which doesn't get parsed (the include is still a DirectiveNode), but
rather a post parse processing step converts the include DirectiveNode's
into IncludeNode's, by parsing the included file. The IncludeNode then
has the included files RootNode in it, along with the included path and
file object. Then, a serializer plugin, when it encounters the
IncludeNode, creates a new context based on the file that the node came
from, and renders the nodes children into that - this ensures that
included files can still have things like snips with paths relative to
the included file.

Doing all this exposed an issue - since we expect included markdown
files to include directives that may contain things that look like
labels, it would have a problem with #149. So I made the snippet parser
be able to turn on or off the filtering out of lines containing labels.
The include directive turns this off by default, while the snip and
other directives turn it on by default. Since that option now existed, I
also added support to all of the above directives to configure it, via a
`filterLabel` flag. This provides a work around, if not a fix, to #149.