Fixed toc handling with includes
This fixes the toc handling so that the headersBelow function takes into
consideration whether a header from an included file is below the toc
directive, and vice versa, if the toc directive is in an included file,
will correctly work out if a header in the including file is below it.
It does this by tracking nested index locations of the includes.

Also fixed the context propagation of rendering included sections so
that included sections can contain a toc.